Rename ErrorMessage.hpp to lerrormessage.hpp for consistency
