+++ Pluto.vcxproj.filters
@@ -71,7 +71,7 @@
     <ClInclude Include="src\ltable.h" />
     <ClInclude Include="src\ljumptab.h" />
     <ClInclude Include="src\lcryptolib.hpp" />
-    <ClInclude Include="src\ErrorMessage.hpp" />
+    <ClInclude Include="src\lerrormessage.hpp" />
     <ClInclude Include="src\ljson.hpp" />
   </ItemGroup>
 </Project>
